change way to check if node is alive
each node now exposes an api with an endpoint with status. The leader now makes a get request to each node which has to control. If the leader can't establish connection with a node, it means the node is down and it needs to be restarted.
Non leader node code removed and replaced with inf. loop.